Allow using a local directory to find boost archives
Instead of downloading boost, allow the user to specify the path to a directory
that contains the required boost archive. To specify this archive directory, use
the new BOOST_ARCHIVE_DIRECTORY cache option. It must point to a valid, existing
directory on the filesystem. To instead download the archives normally, simply
set BOOST_ARCHIVE_DIRECTORY to an empty string or delete it entirely.

When a local directory for archives is being used, any SHA mismatch or missing
archive file will result in permanent failure. In other words, boost will not be
downloaded if these checks fail.